Minor refactoring and cleanup
- merge prototyping and existing safelinks package content
- refactor cmd/usl tool
  - "pull back" code from safelinks package that is specific
    to the cmd/usl tool
- export the cleanURL func as CleanURL to share common URL
  sanitization behavior
- refactor refactor cmd/dsl tool
  - break prototyping.go into separate files
