+++ cmd/usl/main.go
@@ -54,7 +54,7 @@ func main() {
 		inputURLs = input
 
 	default:
-		input, err := safelinks.ProcessInputAsURL(cfg.URL)
+		input, err := ReadURLsFromInput(cfg.URL)
 		if err != nil {
 			fmt.Fprintf(userFeedbackOut, "Failed to parse input as URL: %v\n", err)
 			os.Exit(1)
@@ -63,7 +63,7 @@ func main() {
 		inputURLs = input
 	}
 
-	hasErr := safelinks.ProcessInputURLs(inputURLs, os.Stdout, userFeedbackOut, cfg.Verbose)
+	hasErr := ProcessInputURLs(inputURLs, os.Stdout, userFeedbackOut, cfg.Verbose)
 
 	// Ensure unsuccessful error code if one encountered.
 	if hasErr {
